requirements.txt
numpydoc
pygments
there
rich
jedi
typer
velin
quart-trio
flit
requests
asks
cachetools
